requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
attrs==17.3.0             # via pytest
certifi==2017.11.5        # via requests
chardet==3.0.4            # via requests
dropbox==8.5.0
fancycompleter==0.8       # via pdbpp
idna==2.6                 # via requests
pdbpp==0.9.2
pluggy==0.6.0             # via pytest
py==1.5.2                 # via pytest
pygments==2.2.0           # via pdbpp
pytest==3.3.1
pyyaml==3.12
requests==2.18.4
six==1.11.0               # via dropbox, pytest
urllib3==1.22             # via requests
wmctrl==0.3               # via pdbpp
